requirements.txt
pandas
numpy
sklearn
statsmodel
streamlit